requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
config==0.4.2
dj-database-url==0.5.0
Django==2.0.3
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pep8==1.7.1
psycopg2==2.8.3
pycodestyle==2.5.0
pylint==2.3.1
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
whitenoise==4.1.3
wrapt==1.11.2
